requirements.txt
# BridalVision AI — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────
fastapi==0.111.0
uvicorn==0.29.0

# ── fal.ai SDK ────────────────────────────────────
fal-client==0.4.1

# ── Image Processing ──────────────────────────────
Pillow==10.3.0

# ── Environment Variables ─────────────────────────
python-dotenv==1.0.1

# ── Form Data Support (FastAPI) ───────────────────
python-multipart==0.0.9

# ── Testing ───────────────────────────────────────
pytest==8.2.0
httpx==0.27.0        # For FastAPI test client